FRY-36 Endpoint testing Fixed Bugs. - Renamed url folder to api_urls - Added dummy view to each url
