+++ Fridgify_Backend/api_urls/fridge/fridge_urls.py
@@ -1,12 +1,14 @@
 from django.urls import path, include
 
-from Fridgify_Backend.urls.fridge.content import content_urls
-from Fridgify_Backend.urls.fridge.item import item_urls
-from Fridgify_Backend.urls.fridge.management import management_urls
+from Fridgify_Backend import view
+
+from Fridgify_Backend.api_urls.fridge.content import content_urls
+from Fridgify_Backend.api_urls.fridge.item import item_urls
+from Fridgify_Backend.api_urls.fridge.management import management_urls
 
 urlpatterns = [
     # GET
-    path('', ),
+    path('', view.hello_world),
     # Content Endpoint
     path('content/', include(content_urls)),
     # Item Endpoint
